Fix ChatService test, use real IFileService
