+++ src/vs/workbench/contrib/chat/test/common/chatService.test.ts
@@ -16,6 +16,7 @@ import { IConfigurationService } from '../../../../../platform/configuration/com
 import { TestConfigurationService } from '../../../../../platform/configuration/test/common/testConfigurationService.js';
 import { IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
 import { IEnvironmentService } from '../../../../../platform/environment/common/environment.js';
+import { IFileService } from '../../../../../platform/files/common/files.js';
 import { ServiceCollection } from '../../../../../platform/instantiation/common/serviceCollection.js';
 import { TestInstantiationService } from '../../../../../platform/instantiation/test/common/instantiationServiceMock.js';
 import { MockContextKeyService } from '../../../../../platform/keybinding/test/common/mockKeybindingService.js';
@@ -23,13 +24,14 @@ import { ILogService, NullLogService } from '../../../../../platform/log/common/
 import { IStorageService } from '../../../../../platform/storage/common/storage.js';
 import { ITelemetryService } from '../../../../../platform/telemetry/common/telemetry.js';
 import { NullTelemetryService } from '../../../../../platform/telemetry/common/telemetryUtils.js';
+import { IUserDataProfilesService } from '../../../../../platform/userDataProfile/common/userDataProfile.js';
 import { IWorkspaceContextService } from '../../../../../platform/workspace/common/workspace.js';
 import { IWorkbenchAssignmentService } from '../../../../services/assignment/common/assignmentService.js';
 import { NullWorkbenchAssignmentService } from '../../../../services/assignment/test/common/nullAssignmentService.js';
 import { IExtensionService, nullExtensionDescription } from '../../../../services/extensions/common/extensions.js';
 import { ILifecycleService } from '../../../../services/lifecycle/common/lifecycle.js';
 import { IViewsService } from '../../../../services/views/common/viewsService.js';
-import { mock, TestContextService, TestExtensionService, TestStorageService } from '../../../../test/common/workbenchTestServices.js';
+import { InMemoryTestFileService, mock, TestContextService, TestExtensionService, TestStorageService } from '../../../../test/common/workbenchTestServices.js';
 import { IMcpService } from '../../../mcp/common/mcpTypes.js';
 import { TestMcpService } from '../../../mcp/test/common/testMcpService.js';
 import { ChatAgentService, IChatAgent, IChatAgentData, IChatAgentImplementation, IChatAgentService } from '../../common/chatAgents.js';
@@ -121,8 +123,8 @@ function getAgentData(id: string): IChatAgentData {
 suite('ChatService', () => {
 	const testDisposables = ensureNoDisposablesAreLeakedInTestSuite();
 
-	let storageService: IStorageService;
 	let instantiationService: TestInstantiationService;
+	let testFileService: InMemoryTestFileService;
 
 	let chatAgentService: IChatAgentService;
 
@@ -132,7 +134,7 @@ suite('ChatService', () => {
 			[IWorkbenchAssignmentService, new NullWorkbenchAssignmentService()],
 			[IMcpService, new TestMcpService()],
 		)));
-		instantiationService.stub(IStorageService, storageService = testDisposables.add(new TestStorageService()));
+		instantiationService.stub(IStorageService, testDisposables.add(new TestStorageService()));
 		instantiationService.stub(ILogService, new NullLogService());
 		instantiationService.stub(ITelemetryService, NullTelemetryService);
 		instantiationService.stub(IExtensionService, new TestExtensionService());
@@ -144,11 +146,20 @@ suite('ChatService', () => {
 		instantiationService.stub(IChatService, new MockChatService());
 		instantiationService.stub(IEnvironmentService, { workspaceStorageHome: URI.file('/test/path/to/workspaceStorage') });
 		instantiationService.stub(ILifecycleService, { onWillShutdown: Event.None });
+		instantiationService.stub(IUserDataProfilesService, {
+			defaultProfile: {
+				globalStorageHome: URI.file('/test/path/to/globalStorage')
+			}
+		} as any);
 		instantiationService.stub(IChatEditingService, new class extends mock<IChatEditingService>() {
 			override startOrContinueGlobalEditingSession(): Promise<IChatEditingSession> {
 				return Promise.resolve(Disposable.None as IChatEditingSession);
 			}
 		});
+
+		// Configure test file service with tracking and in-memory storage
+		testFileService = testDisposables.add(new InMemoryTestFileService());
+		instantiationService.stub(IFileService, testFileService);
 
 		chatAgentService = testDisposables.add(instantiationService.createInstance(ChatAgentService));
 		instantiationService.stub(IChatAgentService, chatAgentService);
@@ -173,10 +184,30 @@ suite('ChatService', () => {
 		const session2 = testDisposables.add(testService.startSession(ChatAgentLocation.Chat, CancellationToken.None));
 		session2.addRequest({ parts: [], text: 'request 2' }, { variables: [] }, 0);
 
-		storageService.flush();
+		// Clear sessions to trigger persistence to file service
+		await testService.clearSession(session1.sessionId);
+		await testService.clearSession(session2.sessionId);
+
+		// Verify that sessions were written to the file service
+		assert.strictEqual(testFileService.writeOperations.length, 2, 'Should have written 2 sessions to file service');
+
+		const session1WriteOp = testFileService.writeOperations.find((op: { resource: URI; content: string }) =>
+			op.content.includes('request 1'));
+		const session2WriteOp = testFileService.writeOperations.find((op: { resource: URI; content: string }) =>
+			op.content.includes('request 2'));
+
+		assert.ok(session1WriteOp, 'Session 1 should have been written to file service');
+		assert.ok(session2WriteOp, 'Session 2 should have been written to file service');
+
+		// Create a new service instance to simulate app restart
 		const testService2 = testDisposables.add(instantiationService.createInstance(ChatService));
+
+		// Retrieve sessions and verify they're loaded from file service
 		const retrieved1 = testDisposables.add((await testService2.getOrRestoreSession(session1.sessionId))!);
 		const retrieved2 = testDisposables.add((await testService2.getOrRestoreSession(session2.sessionId))!);
+
+		assert.ok(retrieved1, 'Should retrieve session 1');
+		assert.ok(retrieved2, 'Should retrieve session 2');
 		assert.deepStrictEqual(retrieved1.getRequests()[0]?.message.text, 'request 1');
 		assert.deepStrictEqual(retrieved2.getRequests()[0]?.message.text, 'request 2');
 	});
