+++ src/vs/workbench/contrib/debug/test/browser/mockDebugModel.ts
@@ -6,10 +6,9 @@
 import { DisposableStore } from '../../../../../base/common/lifecycle.js';
 import { NullLogService } from '../../../../../platform/log/common/log.js';
 import { UriIdentityService } from '../../../../../platform/uriIdentity/common/uriIdentityService.js';
+import { TestFileService, TestStorageService } from '../../../../test/common/workbenchTestServices.js';
 import { DebugModel } from '../../common/debugModel.js';
 import { MockDebugStorage } from '../common/mockDebug.js';
-import { TestFileService } from '../../../../test/browser/workbenchTestServices.js';
-import { TestStorageService } from '../../../../test/common/workbenchTestServices.js';
 
 const fileService = new TestFileService();
 export const mockUriIdentityService = new UriIdentityService(fileService);
