+++ src/vs/workbench/services/workingCopy/test/electron-browser/workingCopyBackupTracker.test.ts
@@ -4,47 +4,47 @@
  *--------------------------------------------------------------------------------------------*/
 
 import assert from 'assert';
-import { isMacintosh, isWindows } from '../../../../../base/common/platform.js';
-import { join } from '../../../../../base/common/path.js';
-import { URI } from '../../../../../base/common/uri.js';
-import { hash } from '../../../../../base/common/hash.js';
-import { NativeWorkingCopyBackupTracker } from '../../electron-browser/workingCopyBackupTracker.js';
-import { TextFileEditorModelManager } from '../../../textfile/common/textFileEditorModelManager.js';
-import { IEditorService } from '../../../editor/common/editorService.js';
-import { EditorPart } from '../../../../browser/parts/editor/editorPart.js';
-import { IEditorGroupsService } from '../../../editor/common/editorGroupsService.js';
-import { EditorService } from '../../../editor/browser/editorService.js';
-import { IWorkingCopyBackupService } from '../../common/workingCopyBackup.js';
-import { DisposableStore } from '../../../../../base/common/lifecycle.js';
-import { ensureNoDisposablesAreLeakedInTestSuite, toResource } from '../../../../../base/test/common/utils.js';
-import { IFilesConfigurationService } from '../../../filesConfiguration/common/filesConfigurationService.js';
-import { IWorkingCopyService } from '../../common/workingCopyService.js';
-import { ILogService } from '../../../../../platform/log/common/log.js';
-import { HotExitConfiguration } from '../../../../../platform/files/common/files.js';
-import { ShutdownReason, ILifecycleService } from '../../../lifecycle/common/lifecycle.js';
-import { IFileDialogService, ConfirmResult, IDialogService } from '../../../../../platform/dialogs/common/dialogs.js';
-import { IWorkspaceContextService } from '../../../../../platform/workspace/common/workspace.js';
-import { INativeHostService } from '../../../../../platform/native/common/native.js';
-import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
-import { TestConfigurationService } from '../../../../../platform/configuration/test/common/testConfigurationService.js';
-import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
-import { createEditorPart, registerTestFileEditor, TestBeforeShutdownEvent, TestEnvironmentService, TestFilesConfigurationService, TestFileService, TestTextResourceConfigurationService, workbenchTeardown } from '../../../../test/browser/workbenchTestServices.js';
-import { MockContextKeyService } from '../../../../../platform/keybinding/test/common/mockKeybindingService.js';
-import { IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
-import { IEnvironmentService } from '../../../../../platform/environment/common/environment.js';
-import { TestWorkspace, Workspace } from '../../../../../platform/workspace/test/common/testWorkspace.js';
-import { IProgressService } from '../../../../../platform/progress/common/progress.js';
-import { IWorkingCopyEditorService } from '../../common/workingCopyEditorService.js';
-import { TestContextService, TestMarkerService, TestWorkingCopy } from '../../../../test/common/workbenchTestServices.js';
-import { CancellationToken } from '../../../../../base/common/cancellation.js';
-import { IWorkingCopyBackup, WorkingCopyCapabilities } from '../../common/workingCopy.js';
-import { Event, Emitter } from '../../../../../base/common/event.js';
-import { generateUuid } from '../../../../../base/common/uuid.js';
-import { Schemas } from '../../../../../base/common/network.js';
-import { joinPath } from '../../../../../base/common/resources.js';
 import { VSBuffer } from '../../../../../base/common/buffer.js';
-import { TestServiceAccessor, workbenchInstantiationService } from '../../../../test/electron-browser/workbenchTestServices.js';
+import { CancellationToken } from '../../../../../base/common/cancellation.js';
+import { Emitter, Event } from '../../../../../base/common/event.js';
+import { hash } from '../../../../../base/common/hash.js';
+import { DisposableStore } from '../../../../../base/common/lifecycle.js';
+import { Schemas } from '../../../../../base/common/network.js';
+import { join } from '../../../../../base/common/path.js';
+import { isMacintosh, isWindows } from '../../../../../base/common/platform.js';
+import { joinPath } from '../../../../../base/common/resources.js';
+import { URI } from '../../../../../base/common/uri.js';
+import { generateUuid } from '../../../../../base/common/uuid.js';
+import { ensureNoDisposablesAreLeakedInTestSuite, toResource } from '../../../../../base/test/common/utils.js';
+import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
+import { TestConfigurationService } from '../../../../../platform/configuration/test/common/testConfigurationService.js';
+import { IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
+import { ConfirmResult, IDialogService, IFileDialogService } from '../../../../../platform/dialogs/common/dialogs.js';
+import { IEnvironmentService } from '../../../../../platform/environment/common/environment.js';
+import { HotExitConfiguration } from '../../../../../platform/files/common/files.js';
+import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
+import { MockContextKeyService } from '../../../../../platform/keybinding/test/common/mockKeybindingService.js';
+import { ILogService } from '../../../../../platform/log/common/log.js';
+import { INativeHostService } from '../../../../../platform/native/common/native.js';
+import { IProgressService } from '../../../../../platform/progress/common/progress.js';
 import { UriIdentityService } from '../../../../../platform/uriIdentity/common/uriIdentityService.js';
+import { IWorkspaceContextService } from '../../../../../platform/workspace/common/workspace.js';
+import { TestWorkspace, Workspace } from '../../../../../platform/workspace/test/common/testWorkspace.js';
+import { EditorPart } from '../../../../browser/parts/editor/editorPart.js';
+import { createEditorPart, registerTestFileEditor, TestBeforeShutdownEvent, TestEnvironmentService, TestFilesConfigurationService, TestTextResourceConfigurationService, workbenchTeardown } from '../../../../test/browser/workbenchTestServices.js';
+import { TestContextService, TestFileService, TestMarkerService, TestWorkingCopy } from '../../../../test/common/workbenchTestServices.js';
+import { TestServiceAccessor, workbenchInstantiationService } from '../../../../test/electron-browser/workbenchTestServices.js';
+import { EditorService } from '../../../editor/browser/editorService.js';
+import { IEditorGroupsService } from '../../../editor/common/editorGroupsService.js';
+import { IEditorService } from '../../../editor/common/editorService.js';
+import { IFilesConfigurationService } from '../../../filesConfiguration/common/filesConfigurationService.js';
+import { ILifecycleService, ShutdownReason } from '../../../lifecycle/common/lifecycle.js';
+import { TextFileEditorModelManager } from '../../../textfile/common/textFileEditorModelManager.js';
+import { IWorkingCopyBackup, WorkingCopyCapabilities } from '../../common/workingCopy.js';
+import { IWorkingCopyBackupService } from '../../common/workingCopyBackup.js';
+import { IWorkingCopyEditorService } from '../../common/workingCopyEditorService.js';
+import { IWorkingCopyService } from '../../common/workingCopyService.js';
+import { NativeWorkingCopyBackupTracker } from '../../electron-browser/workingCopyBackupTracker.js';
 
 suite('WorkingCopyBackupTracker (native)', function () {
